+++ bin/tasks/permissions.sh
@@ -4,8 +4,5 @@
 # @Last Modified by:   Roni Laukkarinen
 # @Last Modified time: 2022-01-11 17:16:01
 
-# Let's ensure we have acf-json folder
-mkdir -p ${PROJECT_THEME_PATH}/acf-json
-
 # Ensure file permissions for that folder
-source ${SCRIPTS_LOCATION}/tasks/permissions.sh
+sudo chmod -R 777 ${PROJECT_THEME_PATH}/acf-json