+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Modularize check-updates as task
 * Simplify get-block task by modularizing create-predefined-block as task
+* Ensure file permissions, add permissions task
 
 ### 1.2.2: 2022-09-12